requirements.txt
requests>=2.26.0
pandas>=1.3.0
google-cloud-storage>=2.0.0